[#236] :sparkles: refactor: dto 삭제 -> middleware/validation
