+++ server/src/api/middlewares/validation/auth/loginUser.ts
@@ -1,0 +1,6 @@
+import { body } from 'express-validator';
+
+export default [
+  body('user_id').notEmpty().withMessage('body: user_id가 존재하지 않습니다'),
+  body('password').notEmpty().withMessage('body: password가 존재하지 않습니다'),
+];
