+++ server/src/api/routes/auth.router.ts
@@ -3,7 +3,7 @@ import wrapAsync from '@/utils/wrapAsync';
 import AuthController from '@/controllers/auth.controller';
 import { oauthError } from '@/api/middlewares/oauth';
 import authJWT from '@/api/middlewares/auth.middleware';
-import LoginUserRequest from '@/dtos/auth/loginUser';
+import loginUserValidators from '@/api/middlewares/validation/auth/loginUser';
 import invalidRequest from '@/api/middlewares/invalid-request';
 import { checkPasswordValidators } from '../middlewares/validation/auth/checkPassword';
 
@@ -12,7 +12,7 @@ const router = Router();
 router.get('/', AuthController.gitOAuthUrl);
 router.post(
   '/login',
-  invalidRequest(...LoginUserRequest.validators),
+  invalidRequest(...loginUserValidators),
   wrapAsync(AuthController.Login)
 );
 router.get('/callback', oauthError, wrapAsync(AuthController.callback));
